Move basic functions to utils
